testの分離 #62
test_helper追加
lib/の中で書かれていたテストをtests/test_ に移動
test/unitをminitestに変更
make testでbundle exec make testするよう変更
